Feature/refactor and cleanup

+++ src/services/topics.ts
@@ -2,14 +2,8 @@ import handleErrorMessage from "@/utils/handleErrorMessage";
 import db from "@/db/connection";
 import { schemaTopics } from "@/db/schema/topics";
 import { eq, inArray } from "drizzle-orm";
-import TokenService from "./token";
 import { ApiErrorMessage } from "@/app/api/handleApiError";
-
-// TODO: To be moved out of here
-enum UserRoles {
-  BASIC = 1,
-  ADMIN = 2,
-}
+import UsersService from "./users";
 
 class Topics {
   /** Construct a slug based on the name of the topic */
@@ -69,69 +63,20 @@ class Topics {
    * @param name The name of the topic
    *
    */
-  async create(userId: number, name: string) {
+  async create(name: string) {
     try {
+      const currentUser = await UsersService.getCurrentUser();
+
       await db.insert(schemaTopics).values({
-        name,
+        name: name,
         slug: this.generateUniqueTopicSlug(name),
-        created_by: userId,
+        created_by: currentUser.id,
       });
 
       console.log("TOPICS - Topic created successfully!");
     } catch (error) {
       console.log(
         `TOPICS - Failed creating topic: ${handleErrorMessage(error)}`
-      );
-      throw new Error(handleErrorMessage(error));
-    }
-  }
-
-  /**
-   *
-   * Update the name of an existing topic
-   *
-   * Note: This will be removed, topics will be unique and no changes can be made to them
-   *
-   */
-  async update(id: number, name: string) {
-    try {
-      const tokenDetails = await TokenService.decodeToken();
-      const userID = tokenDetails?.id || 0;
-      const roleID = tokenDetails?.role_id || 1;
-
-      // Find the targeted topic
-      const targetedTopic = await db
-        .select()
-        .from(schemaTopics)
-        .where(eq(schemaTopics.id, id));
-
-      // If topic does not exist
-      if (!targetedTopic[0]) {
-        throw new Error(ApiErrorMessage.NOT_FOUND);
-      }
-
-      // Only allow the user that created the topic (or an admin) to update it
-      if (
-        targetedTopic[0].created_by !== userID ||
-        roleID !== UserRoles.ADMIN
-      ) {
-        throw new Error(ApiErrorMessage.UNAUTHORIZED);
-      }
-
-      // Update the name of the topic
-      await db
-        .update(schemaTopics)
-        .set({ name })
-        .where(eq(schemaTopics.id, id));
-
-      console.log(
-        `TOPICS - Successfully updated the name of topic with ID ${id}.`
-      );
-    } catch (error) {
-      console.log(
-        `TOPICS - Failed updating topic with ID ${id}: ${handleErrorMessage(
-          error
-        )}`
       );
       throw new Error(handleErrorMessage(error));
     }
@@ -146,9 +91,22 @@ class Topics {
    */
   async deleteSpecific(id: number) {
     try {
-      // TODO: Check for topic existance first (?)
+      const targetedTopic = await db
+        .select()
+        .from(schemaTopics)
+        .where(eq(schemaTopics.id, id));
+
+      // Throw error if topic does not exist in the database
+      if (!targetedTopic[0]) throw new Error(ApiErrorMessage.NOT_FOUND);
+
+      // Check for authorization before allowing deletion
+      const currentUser = await UsersService.getCurrentUser();
+      if (targetedTopic[0].created_by !== currentUser.id) {
+        throw new Error(ApiErrorMessage.UNAUTHORIZED);
+      }
+
       await db.delete(schemaTopics).where(eq(schemaTopics.id, id));
-      console.log(`TOPICS - Successfully updated topic with ID ${id}`);
+      console.log(`TOPICS - Successfully deleted topic with ID ${id}`);
     } catch (error) {
       console.log(
         `TOPICS - Failed deleting topic: ${handleErrorMessage(error)}`
@@ -166,6 +124,19 @@ class Topics {
    */
   async deleteBulk(ids: number[]) {
     try {
+      const targetedTopics = await db
+        .select()
+        .from(schemaTopics)
+        .where(inArray(schemaTopics.id, ids));
+
+      if (!targetedTopics.length) throw new Error(ApiErrorMessage.NOT_FOUND);
+
+      // Check if the user that tries deleting the topics is the one that created them
+      const user = await UsersService.getCurrentUser();
+      if (!targetedTopics.every((topic) => topic.created_by === user.id)) {
+        throw new Error(ApiErrorMessage.UNAUTHORIZED);
+      }
+
       await db.delete(schemaTopics).where(inArray(schemaTopics.id, ids));
       console.log(`TOPICS - Successfully removed ${ids.length} topics!`);
     } catch (error) {
